Fix CI failures.
