+++ test/support/src/vfs_helpers.h
@@ -1106,14 +1106,4 @@ class MemFsTest : public VFSTestBase {
   }
 };
 }  // namespace tiledb::test
-namespace Catch {
-template <>
-struct StringMaker<std::pair<std::string, size_t>> {
-  static std::string convert(const std::pair<std::string, size_t>& value) {
-    std::ostringstream oss;
-    oss << "(" << value.first << ", " << value.second << ")";
-    return oss.str();
-  }
-};
-}  // namespace Catch
 #endif
